Amélioration ergonomie

+++ client/src/components/Messages/Message.js
@@ -1,61 +1,35 @@
 import axios from '../../axios.js';
 
-import { FaMinusCircle } from 'react-icons/fa';
-import { FaPlusCircle } from 'react-icons/fa';
 import { FaRegCommentDots } from 'react-icons/fa';
 import { FaRegStar } from 'react-icons/fa';
 import { FaRetweet } from 'react-icons/fa';
 import { FaStar } from 'react-icons/fa';
 import { FaTrashAlt } from 'react-icons/fa';
-import { useEffect, useState } from 'react';
-import { useLocation } from 'react-router-dom';
+import { useState } from 'react';
+import { Link } from 'react-router-dom';
 
 import './styles/message.css';
 
 function Message(props){
-    const [followStatus, setFollowStatus] = useState(false);
     const [starred, setStarred] = useState(false);
-    const location = useLocation();
 
     function handleStarred(evt) {
         setStarred(!starred);
     }
 
-    function follow() {
-        axios.post('/users/follow', { followerLogin: props.currentUserLogin, followedLogin: props.author })
-            .then(res => window.location.reload())
-            .catch(err => console.log(err));
-    }
-
-    function unfollow() {
-        axios.delete('/users/follow', { params: { followerLogin: props.currentUserLogin, followedLogin: props.author } })
-            .then(res => window.location.reload())
-            .catch(err => console.log(err));
-    }
-
-    function isFollower() {
-        axios.get(`/users/isfollower/${props.currentUserLogin}/${props.author}`)
-            .then(res => {
-                setFollowStatus(res.data.found);
-            })
-            .catch(err => console.log(err));
-    }
-
     function deleteMessage() {
-        axios.delete('/messages', { params: { author: props.author, date: props.date, message: props.message, messageId: props.messageId, currentUserLogin: props.currentUserLogin } })
-            .then(res => window.location.reload())
-            .catch(err => console.log(err));
+        if (window.confirm('Souhaitez-vous vraiment supprimer ce message ?')) {
+            axios.delete('/messages', { params: { author: props.author, date: props.date, message: props.message, messageId: props.messageId, currentUserLogin: props.currentUserLogin } })
+                .then(res => window.location.reload())
+                .catch(err => console.log(err));
+        }
     }
-
-    useEffect(() => {
-        isFollower();
-    }, [location]);
 
     return(
         <div id={ props.messageId } className='message'>
             <img draggable='false' src={ props.avatar } alt={ 'Avatar de ' + props.author } />
             <div className='message-metadata'>
-                <h3>{ props.author }</h3>
+                <h3><Link to={ `/profil/${props.author}` }>{ props.author }</ Link></h3>
                 <h4>Le {`${String(props.date.getDate()).padStart(2, '0')}/${String(props.date.getMonth() + 1).padStart(2, '0')}/${props.date.getFullYear()} à ${String(props.date.getHours()).padStart(2, '0')}h${String(props.date.getMinutes()).padStart(2, '0')}`}</h4>
             </div>
             <p>{ props.message }</p>
@@ -69,13 +43,6 @@ function Message(props){
                 <button className='message-features-button'>
                     <FaRegCommentDots title='Commenter' />
                 </button>
-                { props.currentUserLogin !== props.author && (followStatus ?
-                <button className='message-unfollow-button' onClick={ unfollow }>
-                <FaMinusCircle title='Ne plus suivre' />
-                </button> :
-                <button className='message-follow-button' onClick={ follow }>
-                    <FaPlusCircle title='Suivre' />
-                </button>) }
                 { props.currentUserLogin === props.author && 
                 <button className='message-delete-button' onClick={ deleteMessage }>
                     <FaTrashAlt title='Supprimer' />
